app/requirements.txt
python-binance
python-dotenv
mysql-connector-python